+++ pybricksdev/cli/__init__.py
@@ -96,16 +96,17 @@ class Run(Tool):
             choices=["ble", "usb", "ssh"],
         )
         parser.add_argument(
-            "device",
-            metavar="<device>",
-            help="hostname or IP address for SSH connection; "
-            "Bluetooth device name or Bluetooth address for BLE connection; "
-            "serial port name for USB connection",
-        )
-        parser.add_argument(
             "script",
             metavar="<script>",
             help="path to a MicroPython script or inline script",
+        )
+        parser.add_argument(
+            "--name",
+            metavar="<name>",
+            required=False,
+            help="hostname or IP address for SSH connection; "
+            "Bluetooth device name or Bluetooth address for BLE connection; "
+            "serial port name for USB connection",
         )
         parser.add_argument(
             "--wait",
@@ -130,22 +131,31 @@ class Run(Tool):
         # Pick the right connection
         if args.conntype == "ssh":
             # So it's an ev3dev
-            if not validators.ip_address.ipv4(args.device):
+            if args.name is None:
+                print("--name is required for SSH connections", file=sys.stderr)
+                exit(1)
+
+            if not validators.ip_address.ipv4(args.name):
                 raise ValueError("Device must be IP address.")
+
             hub = EV3Connection()
-            device_or_address = args.device
+            device_or_address = args.name
         elif args.conntype == "ble":
             # It is a Pybricks Hub with BLE. Device name or address is given.
             hub = PybricksHub()
-            device_or_address = await find_device(args.device)
-        elif args.conntype == "usb" and args.device == "lego":
+            device_or_address = await find_device(args.name)
+        elif args.conntype == "usb" and args.name == "lego":
             # It's LEGO stock firmware Hub with USB.
             hub = USBRPCConnection()
             device_or_address = "LEGO Technic Large Hub in FS Mode"
         elif args.conntype == "usb":
+            if args.name is None:
+                print("--name is required for USB connections", file=sys.stderr)
+                exit(1)
+
             # It's a Pybricks Hub with USB. Port name is given.
             hub = USBPUPConnection()
-            device_or_address = args.device
+            device_or_address = args.name
         else:
             raise ValueError(f"Unknown connection type: {args.conntype}")